Added run script
