+++ main.py
@@ -1,36 +1,41 @@
-from whot import Whot
+from aiohttp import web
+import asyncio
+from websockets.asyncio.server import serve
+from server import handler
 
-g = Whot(2, number_of_cards=2)
-g.game_mode()
+PORT = 8080
+WEBSOCKET_PORT = 8765
 
-while True:
-    print(g.game_state())
-  
-    print("1. Play")
-    print("2. Market")
-    print("3. Exit")
-    option = int(input("Select option: "))
-    
-    if option == 1:
-        n = int(input("Provide card index: "))
-        result = g.play(n)
-        if result["status"] == "GameOver":
-            print(f"Player: {result['winner']} wins!")
-            break
+# Serve index.html
+async def handle_index(request):
+    return web.FileResponse("index.html")
 
-        if result["status"] == "Request":
-            print("0. Circle")
-            print("1. Square")
-            print("2. Star")
-            print("3. Cross")
-            print("4. Angle")
-            suit = int(input("Provide suit: "))
-            request_card = g.request(suit)
-            print(f"I need: {request_card}")
+# WebSocket server function
+async def websocket_server():
+    async with serve(handler, "0.0.0.0", WEBSOCKET_PORT):
+        print(f"WebSocket server running at ws://0.0.0.0:{WEBSOCKET_PORT}")
+        await asyncio.Future()  # Keep the WebSocket server running
 
-    elif option == 2:
-        g.market()
-    elif option == 3:
-        break
-    else:
-        print("Unknown option")
+# Create an aiohttp web app
+app = web.Application()
+app.router.add_get("/", handle_index)
+app.router.add_static("/", path=".", name="static")
+
+# Function to run the aiohttp server
+async def run_server():
+    runner = web.AppRunner(app)
+    await runner.setup()
+    site = web.TCPSite(runner, "0.0.0.0", PORT)
+    print(f"HTTP server running at http://0.0.0.0:{PORT}")
+    await site.start()
+    await asyncio.Future()  # Keep running
+
+# Run both WebSocket and HTTP server concurrently
+async def main():
+    await asyncio.gather(websocket_server(), run_server())
+
+if __name__ == "__main__":
+    try:
+        asyncio.run(main())
+    except KeyboardInterrupt:
+        print("Server shut down by user.")
